I'm taking my movie break now, got some good progress done

+++ MaizeUI/MZEControlCenterPositionProvider.h
@@ -4,8 +4,8 @@
 	MZELayoutStyle *_layoutStyle;
 	NSArray *_orderedIdentifiers;
 	NSArray *_orderedSizes;
-	NSInteger _numberOfColumns;
-	NSInteger _numberOfRows;
+	CGFloat _numberOfColumns;
+	CGFloat _numberOfRows;
 }
 @property (nonatomic, retain, readwrite) MZELayoutStyle *layoutStyle;
 @property (nonatomic, retain, readwrite) NSArray *orderedIdentifiers;
