+++ MaizeUI/MZELayoutStyle.h
@@ -1,6 +1,6 @@
 @interface MZELayoutStyle : NSObject
-@property (nonatomic, assign) NSUInteger columns;
-@property (nonatomic, assign) NSUInteger rows;
+@property (nonatomic, assign) CGFloat columns;
+@property (nonatomic, assign) CGFloat rows;
 @property (nonatomic, assign) CGFloat inset;
 @property (nonatomic, assign) CGFloat spacing;
 @property (nonatomic, assign) CGFloat moduleSize;
